[DVOT-64] Add first stable version
