+++ components/blocks/mark-down-viewer.tsx
@@ -1,20 +1,11 @@
 import React, { useEffect } from 'react'
 import styled from 'styled-components'
 import MarkDownIt from 'markdown-it'
-import { Md2PdfkitParser } from '@lib/md-2-pdkit-parser'
 
 export const MarkDownViewer = ({ content }) => {
   const wrapperRef = React.useRef(null)
   const md = new MarkDownIt()
   const html = md.render(content || '')
-
-  useEffect(() => {
-
-    if(content) {
-      const parser = new Md2PdfkitParser(content)
-      console.log(parser.getPdfkitOperations())
-    }
-  }, [content])
   
   const handleClickEvent = (event) => {
     if (event.target.tagName === 'A') {
